+++ src/main/java/bc1/gream/domain/gifticon/repository/GifticonRepositoryCustomImpl.java
@@ -1,7 +1,7 @@
 package bc1.gream.domain.gifticon.repository;
 
 import bc1.gream.domain.gifticon.entity.Gifticon;
-import bc1.gream.domain.order.entity.QGifticon;
+import bc1.gream.domain.gifticon.entity.QGifticon;
 import bc1.gream.domain.order.entity.QOrder;
 import bc1.gream.domain.user.entity.User;
 import com.querydsl.jpa.impl.JPAQueryFactory;
